Adjusted code indentation by using spaces instead of tabs.
- Replaced tabs with spaces for indentation.
- Adapted `.clang-format` and `.editorconfig` files.
- Reformatted some comments.
- Disabled automatic indentation for certain blocks.
- Fixed style of TODO comments.

+++ src/util/parser_util.c
@@ -26,57 +26,58 @@
 #include <tss2/tss2_tpm2_types.h>
 
 CHARRA_RC parse_pcr_value(char* start, char* eol, uint8_t* pcr_value) {
-	// search for the start of the PCR hex value
-	char* hex_start;
-	for (hex_start = start;
-		 !(*(hex_start - 2) == '0' && *(hex_start - 1) == 'x'); hex_start++) {
-		if (hex_start >= eol) {
-			return CHARRA_RC_ERROR;
-		}
-	} // loop ends on first character after the '0x'
+    // search for the start of the PCR hex value
+    char* hex_start;
+    for (hex_start = start;
+            !(*(hex_start - 2) == '0' && *(hex_start - 1) == 'x');
+            hex_start++) {
+        if (hex_start >= eol) {
+            return CHARRA_RC_ERROR;
+        }
+    }  // loop ends on first character after the '0x'
 
-	// iterate over all bytes of the digest
-	for (uint32_t digest_index = 0; digest_index < TPM2_SHA256_DIGEST_SIZE;
-		 digest_index++) {
-		// hex_index is the byte in string representation at the
-		// current digest_index
-		char* hex_index = hex_start + (digest_index * 2);
-		if (hex_index + 1 >= eol) {
-			return CHARRA_RC_ERROR;
-		}
+    // iterate over all bytes of the digest
+    for (uint32_t digest_index = 0; digest_index < TPM2_SHA256_DIGEST_SIZE;
+            digest_index++) {
+        // hex_index is the byte in string representation at the
+        // current digest_index
+        char* hex_index = hex_start + (digest_index * 2);
+        if (hex_index + 1 >= eol) {
+            return CHARRA_RC_ERROR;
+        }
 
-		// convert byte in string representation to byte as uint8_t
-		char byte_as_string[3] = {0};
-		// copy substring into other string because otherwise strtoul
-		// would read more than one byte
-		memcpy(byte_as_string, hex_index, 2);
-		byte_as_string[2] = '\0';
-		errno = 0;
-		char* eol = NULL;
-		unsigned long int hex_value = strtoul(byte_as_string, &eol, 16);
-		if (eol == byte_as_string || errno != 0 || hex_value > 255) {
-			return CHARRA_RC_ERROR;
-		}
-		pcr_value[digest_index] = (uint8_t)hex_value;
-	}
-	return CHARRA_RC_SUCCESS;
+        // convert byte in string representation to byte as uint8_t
+        char byte_as_string[3] = {0};
+        // copy substring into other string because otherwise strtoul
+        // would read more than one byte
+        memcpy(byte_as_string, hex_index, 2);
+        byte_as_string[2] = '\0';
+        errno = 0;
+        char* eol = NULL;
+        uint32_t hex_value = strtoul(byte_as_string, &eol, 16);
+        if (eol == byte_as_string || errno != 0 || hex_value > 255) {
+            return CHARRA_RC_ERROR;
+        }
+        pcr_value[digest_index] = (uint8_t)hex_value;
+    }
+    return CHARRA_RC_SUCCESS;
 }
 
 char* find_end_of_line(char* start, char* end) {
-	for (char* c = start; c < end; c++) {
-		if (*c == '\n') {
-			return c;
-		}
-	} // loop ends when end of file or end of line is reached
-	return end;
+    for (char* c = start; c < end; c++) {
+        if (*c == '\n') {
+            return c;
+        }
+    }  // loop ends when end of file or end of line is reached
+    return end;
 }
 
 int parse_pcr_index(char* index_start) {
-	errno = 0;
-	char* end = NULL;
-	int pcr_index = strtoul(index_start, &end, 10); // parse digits as index
-	if (end == index_start || errno != 0 || pcr_index >= TPM2_MAX_PCRS) {
-		return -1;
-	}
-	return pcr_index;
+    errno = 0;
+    char* end = NULL;
+    int pcr_index = strtoul(index_start, &end, 10);  // parse digits as index
+    if (end == index_start || errno != 0 || pcr_index >= TPM2_MAX_PCRS) {
+        return -1;
+    }
+    return pcr_index;
 }
